Avoid providing a Streamer interface.
The benefits of this are fewer lines of code and less duplication of
method signatures. This coding style is also consistent with that
adopted by much of the  Go standard library.

The main disadvantage is that coupling is increased (since the client
of streamer refers to Streamer's concrete type). Also, the public
interface of Streamer is possibly less clear now that it is scattered
throughout the implementation code rather than being gathered together
along with its documentation.

For an internal helper type, the above advantages just about outweigh
the disadvantages.

[#100629090]